pyproject.toml
[build-system]
requires = ["uv_build >= 0.9.5, <0.10.0"]
build-backend = "uv_build"

[project]
name = "final-project-demo"
version = "0.1.0"
description = "YouTube trending video analysis package — merges Kaggle data with YouTube API for EDA and predictive modeling"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Summer Price", email = "summeraskey@gmail.com" },
    { name = "Jane Gustafson" }
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent"
]
dependencies = [
    "python-dotenv>=1.0.0",
    "google-api-python-client>=2.194.0",
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "kagglehub>=1.0.0",
    "matplotlib>=3.10.8",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "streamlit>=1.40.0",
    "build>=1.4.3",
    "twine>=6.2.0",
]
[tool.uv_build]
exclude = [
    "src/final_project_demo/data",
    "src/final_project_demo/*.png",
    "src/final_project_demo/*.ipynb",
]